init: member 모듈 구조 수정, entity 컨벤션 적용
