--- src/main/java/com/samill/missionary_backend/member/admin/domain/dto/TokenDto.java
@@ -1,17 +1,0 @@
-package com.samill.missionary_backend.member.admin.domain.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class TokenDto {
-
-    private String token;
-}
