--- src/main/java/com/samill/missionary_backend/member/admin/domain/service/AdminService.java
@@ -1,5 +1,0 @@
-package com.samill.missionary_backend.member.admin.domain.service;
-
-public interface AdminService {
-
-}
