+++ src/main/java/com/samill/missionary_backend/member/admin/dto/UserDto.java
@@ -1,0 +1,27 @@
+package com.samill.missionary_backend.member.admin.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class UserDto {
+
+    @NotNull
+    private String userId;
+
+    @JsonProperty(access = JsonProperty.Access.WRITE_ONLY)
+    @NotNull
+    private String password;
+
+    @NotNull
+    private String username;
+}
