--- src/main/java/com/samill/missionary_backend/member/admin/infrastructure/service/AdminServiceImpl.java
@@ -1,6 +1,0 @@
-package com.samill.missionary_backend.member.admin.infrastructure.service;
-
-import com.samill.missionary_backend.member.admin.domain.service.AdminService;
-
-public class AdminServiceImpl implements AdminService {
-}
