+++ src/main/java/com/samill/missionary_backend/member/admin/service/AdminService.java
@@ -1,0 +1,6 @@
+package com.samill.missionary_backend.member.admin.service;
+
+import com.samill.missionary_backend.member.admin.service.AdminService;
+
+public class AdminService {
+}
