--- src/main/java/com/samill/missionary_backend/member/member/domain/dto/LoginDto.java
@@ -1,22 +1,0 @@
-package com.samill.missionary_backend.member.member.domain.dto;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class LoginDto {
-
-    @NotNull
-    private String username;
-
-    @NotNull
-    private String password;
-}
