--- src/main/java/com/samill/missionary_backend/member/member/domain/service/MemberService.java
@@ -1,5 +1,0 @@
-package com.samill.missionary_backend.member.member.domain.service;
-
-public interface MemberService {
-
-}
