--- src/main/java/com/samill/missionary_backend/member/member/infrastructure/service/MemberServiceImpl.java
@@ -1,6 +1,0 @@
-package com.samill.missionary_backend.member.member.infrastructure.service;
-
-import com.samill.missionary_backend.member.member.domain.service.MemberService;
-
-public class MemberServiceImpl implements MemberService {
-}
