+++ src/main/java/com/samill/missionary_backend/member/member/service/MemberService.java
@@ -1,0 +1,4 @@
+package com.samill.missionary_backend.member.member.service;
+
+public class MemberService {
+}
