--- src/main/java/com/samill/missionary_backend/member/user/domain/service/UserService.java
@@ -1,5 +1,0 @@
-package com.samill.missionary_backend.member.user.domain.service;
-
-public interface UserService {
-
-}
