+++ src/main/java/com/samill/missionary_backend/member/user/entity/User.java
@@ -1,0 +1,50 @@
+package com.samill.missionary_backend.member.user.entity;
+
+import com.samill.missionary_backend.common.entity.BaseEntity;
+import com.samill.missionary_backend.member.member.entity.Member;
+import jakarta.persistence.*;
+import lombok.*;
+import org.hibernate.annotations.GenericGenerator;
+import org.hibernate.annotations.SQLDelete;
+import org.hibernate.annotations.SQLRestriction;
+import org.hibernate.annotations.Where;
+
+import java.time.OffsetDateTime;
+
+@Builder
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+@SQLDelete(sql = "UPDATE app_user SET deleted_at = current_timestamp WHERE id = ?")
+@SQLRestriction(value = "deleted_at is NULL")
+@Table(name = "app_user")
+public class User extends BaseEntity {
+
+    @Id
+    @GeneratedValue(generator = "uuid2")
+    @GenericGenerator(name = "uuid2", strategy = "uuid2")
+    private String id;
+    private String loginId;
+    private String name;
+    private String identityNumber;
+    private String phoneNumber;
+    private String birthDate;
+    private String gender;
+    private String password;
+    private String email;
+    //세례 여부
+    @Builder.Default
+    private Boolean is_baptized = Boolean.FALSE;
+
+    //세례 일시
+    private OffsetDateTime baptizedAt;
+
+    @OneToOne
+    @JoinColumn(name = "member_id", referencedColumnName = "memberId")
+    private Member member;
+
+    private OffsetDateTime deletedAt;
+
+}
