--- src/main/java/com/samill/missionary_backend/member/user/infrastructure/service/UserServiceImpl.java
@@ -1,6 +1,0 @@
-package com.samill.missionary_backend.member.user.infrastructure.service;
-
-import com.samill.missionary_backend.member.user.domain.service.UserService;
-
-public class UserServiceImpl implements UserService {
-}
