+++ src/main/java/com/samill/missionary_backend/member/user/service/UserService.java
@@ -1,0 +1,4 @@
+package com.samill.missionary_backend.member.user.service;
+
+public class UserService {
+}
