--- src/main/java/com/samill/missionary_backend/terms/domain/service/TermsService.java
@@ -1,5 +1,0 @@
-package com.samill.missionary_backend.terms.domain.service;
-
-public interface TermsService {
-
-}
