--- src/main/java/com/samill/missionary_backend/terms/infrastructure/service/TermsServiceImpl.java
@@ -1,6 +1,0 @@
-package com.samill.missionary_backend.terms.infrastructure.service;
-
-import com.samill.missionary_backend.terms.domain.service.TermsService;
-
-public class TermsServiceImpl implements TermsService {
-}
