+++ src/main/java/com/samill/missionary_backend/terms/service/TermsService.java
@@ -1,0 +1,4 @@
+package com.samill.missionary_backend.terms.service;
+
+public class TermsService{
+}
